browser icon bug and widget-folder bug
the default Application function, has problems locating apps in flatpak, so code was added to avoid this error, when the script is presented it will avoid using said icon in the dock.

some users have an error in the automatic creation of the folders, that is, the dekstop folders, images, music, etc.
the js script contains the addition of the "org.kde.plasma.folder" widget which shows a specific folder, in the script it is stipulated that the folder is downloads,
However, when this folder does not exist, an inconsistent icon is generated, so now the script first checks for the existence of said folder, otherwise no folder was defined.